Added in wildcards + should work on all widgets
Added getRawFeedback() to ToyWidget which returns the actual feedback
value for that widget (whereas getFeedback() returns the value that
will be actioned (with wildcards etc processed, or the parent’s if
relevant),

When grid is created tell each widget which row, column and index it is
(zero based)

Added wildcard processing:
{r} will be replaced with the (zero based) row of the widget
{r1} will be replaced with the (one based) row of the widget
{c} will be replaced with the (zero based) column of the widget
{c1} will be replaced with the (one based) column of the widget
{n} will be replaced with the (zero based) index of the widget
{n1} will be replaced with the (one based) index of the widget

I also made the default feedback value of a widget /XX/{n1} for
testing.. begs the question what it should be.

A few things todo/to consider

- Extend the above to work with all OSC fields - extract the wildcard
code so it can be re-used
- Consider default values - maybe default so that Grid has the default
(/XXX/{n}), make the placeholderText on the field in the edit box into
the parent’s value
- Is there a way to also inherit colours etc? Don’t think these would
use any wildcards but would be useful. To complicate matters it would
be good to be able to edit these on a per-row. Well, actually it would
be great if all of this could be per grid, per row or per cell…
- Would be nice to also have an “offset” which can be modified. This
would mean the user can somehow have up/down buttons which modify a
base value - then we could have the wildcards of {oc}, {or}, {on} etc
where these are the cell/column/num plus an offset. We would probably
want a different offset for each one. This would allow the user to
cycle through presets/palletes etc in an eos scenario